Update index.html + index.css + relog.css + task_handler.js + reglog_handler.js + 2 images + main.js
1. Added session maintanence code:
If a user closes the application without logging out then
next time the user will be logged in automatically.
Uses a file (last_login.txt) for storing the user ID of the last
logged in user. On manual log-out the file is truncated.

Following methods are used for the above purpose:
loadLastLoginSession()
addLastLoginSession()
removeLastLoginSession()

2. Added a background wallpaper.
3. User's full name is split and then the first letter in each split part is
concatenated.
4. Header color changes to purple when no user is logged in and
to red when logged in. Buttons are changed accordingly. Task
input box is also modified accordingly.

Note: developer tools options set to false for 0.0.2