Update project_handler.js + task_handler.js + index_main.js
1. A little refactoring showProjectTaskListCont() is used
to show the project task list container.

2. Remove the project and project task from the global array PROJECT_LIST and
PROJECT_TASK_LIST too when a task is deleted from record. Since during the same
time, if the deleted thing is added (try) again, before log out or closing window, then
the input box will show thing to be already added since it has not been removed the
main holders, so it is required that the thing should be removed from the global
arrays too.

3. showMainTaskListCont() for showing the main task container in index_main.js. Some refactoring here too.