Update reglog_handler.js
1. registerUser(): Elaborated the if-else cases at the start for following scenario,
1.1. If 'data-store' directory is not present then add the 'data-store' directory
along with the 'next_user_id.txt' file inside it.

1.2. If 'next_user_id.txt' is not present then add the 'next_user_id.txt' file and initialize the NEXT_USER_ID variable with 0.

1.3. Else read the next_user_id.txt's content.

2. loginUser():
2.1 If 'user_list.txt' is not present then add it and invoke the callback() with false status.

2.2 If password does not matches then return false status along with 'WRONG_PASSWORD'
error statement.

2.3 Added one missing callback() at line 143. If user_list.txt file returned empty string then the
for loop wouldn't have started thus leading no callback() being called by loginUser() function.
So added a callback() just after for() loop in case for loop does not execute, callback() will be
executed.

3. When the 'login-btn-badge' is shown with a message, it previously
showed one type of error for any false status, that is, 'user not found'.
This is a problem, if the status is false because of wrong password, so
had to add an if-else check for 'result.error' parameter's value to show
the right error message.

4. loadLastLoginSession(): Initialize the last_login.txt file (in case it
does not exists) with 0 not empty string.

5. Removed the console.log() and logger.warn/info()'s wherever present.