Update reglog_handler.js
line 7 - 26: Added more cases of file or directory missing and what to
do about it.

line 55: SESSION_STORE variable wasn't initialized for the first time (during registeration).

line 184: addLastLoginSession() needs userId as a parameter, undefined will go in the file, if no parameter is passed.

line 321: correcting the false mistake of initializing the file with
a userId. 0 counts as the first registered user's ID.

line 327: In case of error, if file contains undefined. This can happen
if SESSION_STORE variable is undefined, then undefined will be
written into the file in addLastLoginSession(), and while reading the file undefined will be read as a string not the special undefined
object, so need to compare the file-read value with a string written
as "undefined".