Update index_main.js + task_handler.js + project_handler.js
Added a little enhancement for showing the task's progress.
When task completes, a green border is added.
For uncompleted tasks whose time is over red outline is added.
For this 'status' attribute in the <li> attribute is added which tells
whether the task is complete or not.

Refactored the two separate click listeners on the different types of
task lists into one and placed the combined function into index_main.js