Update key_handler.js
Added a fix when ctrl + enter is pressed.
If nothing is entered in the input then a task will be created with blank text.
We don't want that.
Same for project naming also.

Corrected the task input dropdown functioning, if task input is already open
and if project input is tried to open then it will also open below it and vice-versa
and that's very wrong. So before dropping any box, lift up previous one.