Update porting.js
JSON for ill-formed and in-complete.

'taskId' was referring to taskComplete property value.
Added 'taskComplete' property.

'projects' was not correctly formed. Contained associative object
inside the array. Had to explicitly fetch the value of the key in the
array and place those values in a separate numeral indexed array.

projectTasks JSON was also ill-formed.
Added 'userPass" in the 'userInfo' JSON.