Update server.js
1. /poll - When profiles comes check if the user exists otherwise, the
same userId will be appended again and again in the user_list.txt file.
On each request arrival sync the profiles with other servers too.

2. /sync - Used for receiving requests from other servers when
synchronization of profiles's updated details is required. A
requesting server will send the details (contain the profiles's data) and
the receiving server will compare the portingTime (timestamp) of the
already stored information with the arrived information, if stored
information is stale it will be updated with the new details
(OBVIOUSLY).