Update reglog_handler.js
In searchUser(): pop out the last empty value in 'userList []'.

Click listener for changing the login email.

When valuing the email tag element, change-login-email icon element is
removed, so had to add <p> element to show the text in that element.

mainCallback is the name of the final callback in searchUser() with
respect to sub callbacks in parallel functions.