Update index_main.js
'servers' array containing the URL's should only contain the address
and port and with an ending slash.
Endpoint should be added when the URL required. 'response.url' is
replaced with serverUrl.
Show/hide the perform-sync icon when online/offline.

Added a click listener on the perform-sync icon.
First, fetch the profile information from the server (central server
approach is used for now), and then that information is used to update
the profile.